Updated home File
Added an "Upgrade User" button that can be used by Guest users to update their accounts to full-blown user accounts.

Added a modal that will be used to edit and update any of the print elements on the page. I'm not quite sure how this will look just yet, since I still need to update the CSS, but I believe that this is all the code I will need for these elements.