Updated main File
Updated main Javascript file with onload function, initialization functions, and a special function to check the variables on each page to ensure they are all allocated properly. I did this latter function because the sheer size of this single file would make it extremely difficult to verify the integrity of the necessary variables on each page.

I also updated each variable to use "let" instead of "var".