Updated main File
Updated main Javascript file with the last commit's "more specific title ID's" as well as booleans to help ensure the proper loading of each page.

The wording of "Turn Off AdBlockers" is not to imply that I plan on using an advertisements on this website (I don't), but rather to prevent against accidental element removal which may cause errors later on.