Updated main (CSS) File
Updated main CSS file to make all the elements look pretty! As per usual, this CSS is pretty much identical to the Gifty code. For the time being this is the way I want it to be, but I may have to adjust this later because of the modals for prints, users, and filaments.