Updated main File
Added a directory to show the general location of each function as well as sections that are significant to the program. As functions get added, my intention is to update the directory accordingly. To help keep things straightforward, I already added some basic functions that I will need to operate each page and placed them appropriately in each section.